refactor: add scroll_menu.dart file
