+++ lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'src/scroll_menu.dart';
+
 void main() => runApp(new MyApp());
 
 class MyApp extends StatelessWidget {
@@ -16,19 +18,14 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatelessWidget {
-
   @override
-    Widget build(BuildContext context) {
-      // TODO: implement build
-      return Scaffold(
+  Widget build(BuildContext context) {
+    return Scaffold(
         appBar: AppBar(
           title: Text("circle"),
         ),
-        body: Center(
-          child: CircleImages()
-          )
-      );
-    }
+        body: Center(child: CircleImages()));
+  }
 }
 
 // class MyHomePage extends StatelessWidget {
@@ -42,7 +39,7 @@ class MyHomePage extends StatelessWidget {
 //         ),
 //         body: Center(
 //           child: Container(
-                  
+
 //                   height: 80.0,
 //                   width: 80.0,
 //                   margin: EdgeInsets.only(
@@ -61,306 +58,31 @@ class MyHomePage extends StatelessWidget {
 //                       image: NetworkImage("https://cdn.dribbble.com/users/1368/screenshots/1785863/icons_2x.png")
 //                     )
 //                   )
-//                 ) 
+//                 )
 //         ) ,
 //       );
 //     }
 // }
 
-class CircleImages extends StatefulWidget{
-  
-@override
-  State<StatefulWidget> createState() {
-    // TODO: implement createState
-    return CircleWidgets();
-  }
-}
+//  class BodyLayout extends StatelessWidget {
+//     @override
+//     Widget build(BuildContext context) {
+//       return _myListView(context);
+//     }
+//   }
 
-class CircleWidgets extends State<CircleImages>{
-  @override
-    Widget build(BuildContext context) {
-      List<Widget> widgets = [];
-      // for(var x = 0 ; x < 10 ; x++){
-        widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-             widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-             widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-             widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-             widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-            widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-            widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-            widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-            widgets.add(Container(
-                  height: 60.0,
-                  width: 60.0,
-                  margin: EdgeInsets.all(
-                    6.0
-                  ),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(100.0),
-                    boxShadow:[
-                      new BoxShadow(
-                        color: Color.fromARGB(100, 0, 0, 0),
-                        blurRadius: 5.0,
-                        offset: Offset(5.0, 5.0)
-                      )
-                    ], 
-                    border: Border.all(
-                      width: 2.0,
-                      style:BorderStyle.solid ,
-                      color: Color.fromARGB(255, 0 , 0, 0)
-                    ),
-                    image: DecorationImage(
-                      fit: BoxFit.cover,
-                      image: NetworkImage("https://upload.wikimedia.org/wikipedia/commons/thumb/e/e6/Noto_Emoji_KitKat_263a.svg/1200px-Noto_Emoji_KitKat_263a.svg.png")
-                    )
-                  )
-                ) 
-            );
-      // }
-      return Container(
-        height: 80.0,
-
-        child: ListView(
-          scrollDirection: Axis.horizontal,
-          padding: EdgeInsets.all(2.0),
-          children: widgets
-        )
-      );
-      
-    }
-}
-  
-
-
-  //  class BodyLayout extends StatelessWidget {
-  //     @override
-  //     Widget build(BuildContext context) {
-  //       return _myListView(context);
-  //     }
-  //   }
-
-
-    // Widget _myListView(BuildContext context) {
-    //   return ListView.builder(
-    //     scrollDirection: Axis.horizontal,
-    //     itemBuilder: (context, index) {
-    //       return Container(
-    //         margin: const EdgeInsets.symmetric(horizontal: 1.0),
-    //         color: Colors.tealAccent,
-    //         child: Text('$index'),
-    //       );
-    //     },
-    //   );
-    // }
-
+// Widget _myListView(BuildContext context) {
+//   return ListView.builder(
+//     scrollDirection: Axis.horizontal,
+//     itemBuilder: (context, index) {
+//       return Container(
+//         margin: const EdgeInsets.symmetric(horizontal: 1.0),
+//         color: Colors.tealAccent,
+//         child: Text('$index'),
+//       );
+//     },
+//   );
+// }
 
 // class MyHomePage extends StatefulWidget {
 //   MyHomePage({Key key, this.title}) : super(key: key);
@@ -562,4 +284,3 @@ class CircleWidgets extends State<CircleImages>{
 
 // typedef void OnPickImageCallback(
 //     double maxWidth, double maxHeight, int quality);
-
